refactor(stock-module): simplify StockKlineChart by utilizing helper functions
- Removed the isLightTheme function and integrated its logic into the buildKlineOptions function for cleaner code.
- Introduced useKlineRange hook to manage range selection, enhancing the chart's functionality and maintainability.
- Updated the options generation for the StockKlineChart to leverage the new helper functions, improving readability and performance.
